move shareable code from Flink runner to JobServerDriver
