+++ runners/java-fn-execution/build.gradle
@@ -32,6 +32,7 @@ dependencies {
   shadow project(path: ":beam-vendor-sdks-java-extensions-protobuf", configuration: "shadow")
   shadow library.java.vendored_grpc_1_13_1
   shadow library.java.slf4j_api
+  shadow library.java.args4j
   testCompile project(":beam-sdks-java-harness")
   testCompile project(path: ":beam-runners-core-construction-java", configuration: "shadow")
   testCompile project(path: ":beam-runners-core-java", configuration: "shadowTest")
